sam_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
transformers==4.47.0
torch>=2.3.0
Pillow==11.0.0
python-multipart==0.0.17

